Fixed broken tests

+++ tests/test_Rational.cpp
@@ -6,8 +6,8 @@ TEST(test_Rational, constructors) {
   // Test creation of Rational numbers
   Rational a = Rational(  5,6);
   Rational b = Rational(-17,12);
-  ASSERT_EQ(a, +5.0/6.0);
-  ASSERT_EQ(b,-17.0/12.0);
+  ASSERT_EQ(double(a), +5.0/6.0);
+  ASSERT_EQ(double(b),-17.0/12.0);
 } /* TEST(test_Rational, constructors) */
 
 TEST(test_Rational, arithmetic) {
@@ -15,15 +15,15 @@ TEST(test_Rational, arithmetic) {
   Rational b = Rational(-17,12);
   Integer  i = 5;
   // Test basic arithmetic operations between two Rational numbers
-  ASSERT_EQ(a+b, -7.0/12.0);
-  ASSERT_EQ(a-b,+27.0/12.0);
+  ASSERT_EQ(double(a+b), -7.0/12.0);
+  ASSERT_EQ(double(a-b),+27.0/12.0);
   // Test basic arithemtic operations between a Rational number and an Integer
-  ASSERT_EQ(a*i,  (5*i)/6.0 );
-  ASSERT_EQ(b*i,(-17*i)/12.0);
-  ASSERT_EQ(a/i,  (5/i)/6.0 );
-  ASSERT_EQ(b/i,(-17/i)/12.0);
-  ASSERT_EQ(a%i,  (5%i)/6.0 );
-  ASSERT_EQ(b%i,(-17%i)/12.0);
+  ASSERT_EQ(double(a*i),  (5*i)/6.0 );
+  ASSERT_EQ(double(b*i),(-17*i)/12.0);
+  ASSERT_EQ(double(a/i),  (5/i)/6.0 );
+  ASSERT_EQ(double(b/i),(-17/i)/12.0);
+  ASSERT_EQ(double(a%i),  (5%i)/6.0 );
+  ASSERT_EQ(double(b%i),(-17%i)/12.0);
 } /* TEST(test_Rational, arithmetic) */
 
 TEST(test_Rational, comparison) {
